riscv/syscalls: Allow syscalls to return capabilities
This is a RISC-V port of two Morello changes:

    227f0d5f8 ("arm64/syscalls: Allow syscalls to return capabilities")

    Tweak the syscalls framework to allow playing with the actual syscall
    prototypes, enabling selectively marking certain syscalls as special cases,
    where the type of expected return value differs from currently default _long_
    type. The definition of syscall table gets slightly extended to accommodate
    additional syscall function prototypes that can be distinguished in runtime.
    Additionally invoke_syscall gets aware of potentially different entries in
    the syscall table (aka syscall fn) enabling it to determine the right
    function type to be used for actually syscall along with proper handling of
    the syscall's return value.


and

    196b8f232 ("arm64: syscall: Allow syscalls without parameters to return capabilities")

    Before this change, only syscalls with parameters were allowed to
    return capabilities. This was previously done by modifying the
    __SYSCALL_DEFINEx macro, which doesn't handle SYSCALL_DEFINE0.
    The SYSCALL_DEFINE0 macros is tweaked to allow the change of the
    return type.


Note that I've not ported across the changes to sys_call_table[] which
adds a flag indicated whether a function returns a capability. At first
glance this doesn't appear to be required.

+++ arch/riscv/include/asm/syscall.h
@@ -97,7 +97,7 @@ static inline int syscall_get_arch(struct task_struct *task)
 #endif
 }
 
-typedef long (*syscall_t)(const struct pt_regs *);
+typedef uintptr_t (*syscall_t)(const struct pt_regs *);
 static inline void syscall_handler(struct pt_regs *regs, ulong syscall)
 {
 	syscall_t fn;
